refactor((reorder MARC record fields)):
refactoring

Closes: MODQM-407
